fix: updated dependencies
fix: functions test cover issue
fix: seo.service and app.component.spec have split because of new tslint rules
fix: a few new tslint issues
