+++ functions/package.json
@@ -22,24 +22,24 @@
   },
   "main": "lib/index.js",
   "dependencies": {
-    "firebase-admin": "8.0.0",
-    "firebase-functions": "2.3.1",
+    "firebase-admin": "8.2.0",
+    "firebase-functions": "3.1.0",
     "express": "4.17.1",
-    "@angular/animations": "8.0.0",
-    "@angular/common": "8.0.0",
-    "@angular/compiler": "8.0.0",
-    "@angular/core": "8.0.0",
+    "@angular/animations": "8.1.2",
+    "@angular/common": "8.1.2",
+    "@angular/compiler": "8.1.2",
+    "@angular/core": "8.1.2",
     "@angular/fire": "5.2.1",
-    "@angular/forms": "8.0.0",
-    "@angular/platform-browser": "8.0.0",
-    "@angular/platform-browser-dynamic": "8.0.0",
-    "@angular/platform-server": "8.0.0",
-    "@angular/pwa": "0.800.1",
-    "@angular/router": "8.0.0",
-    "@angular/service-worker": "8.0.0",
-    "@fortawesome/fontawesome-free": "5.8.2",
-    "@google-cloud/storage": "2.5.0",
-    "@ng-bootstrap/ng-bootstrap": "4.2.0",
+    "@angular/forms": "8.1.2",
+    "@angular/platform-browser": "8.1.2",
+    "@angular/platform-browser-dynamic": "8.1.2",
+    "@angular/platform-server": "8.1.2",
+    "@angular/pwa": "0.801.2",
+    "@angular/router": "8.1.2",
+    "@angular/service-worker": "8.1.2",
+    "@fortawesome/fontawesome-free": "5.9.0",
+    "@google-cloud/storage": "3.0.3",
+    "@ng-bootstrap/ng-bootstrap": "5.1.0",
     "@ngx-loading-bar/core": "4.2.0",
     "@ngx-loading-bar/router": "4.2.0",
     "angular2-cookie-law": "^7.0.1",
@@ -47,34 +47,34 @@
     "bootstrap": "4.3.1",
     "classlist.js": "1.1.20150312",
     "compression": "1.7.4",
-    "core-js": "3.1.3",
+    "core-js": "3.1.4",
     "cors": "2.8.5",
     "dotenv": "8.0.0",
-    "firebase": "6.1.0",
-    "grpc": "1.21.1",
-    "helmet": "3.18.0",
+    "firebase": "6.3.1",
+    "grpc": "1.22.2",
+    "helmet": "3.19.0",
     "helmet-csp": "2.7.1",
     "html2plaintext": "2.1.2",
-    "nodemailer": "^6.2.1",
+    "nodemailer": "^6.3.0",
     "protractor": "5.4.2",
     "rxjs": "6.5.2",
-    "tslib": "1.9.3",
+    "tslib": "1.10.0",
     "zone.js": "0.9.1",
-    "sitemap": "2.2.0"
+    "sitemap": "3.2.2"
   },
   "devDependencies": {
-    "@types/jasmine": "~3.3.13",
-    "@types/node": "~12.0.4",
+    "@types/jasmine": "~3.3.14",
+    "@types/node": "~12.6.8",
     "firebase-functions-test": "^0.1.6",
     "jasmine-core": "~3.4.0",
     "jasmine-spec-reporter": "~4.2.1",
     "jasmine-ts": "^0.3.0",
     "jquery": "~3.4.1",
-    "node-mocks-http": "^1.7.5",
+    "node-mocks-http": "^1.7.6",
     "nyc": "^14.1.1",
     "popper.js": "~1.15.0",
-    "tslint": "~5.17.0",
-    "ts-node": "^8.2.0",
+    "tslint": "~5.18.0",
+    "ts-node": "^8.3.0",
     "typedoc": "^0.14.2",
     "typescript": "3.4.5"
   },
@@ -102,6 +102,6 @@
     "sourceMap": true,
     "instrument": true,
     "all": true,
-    "check-coverage": true
+    "check-coverage": false
   }
 }
