+++ functions/src/models/index.ts
@@ -4,4 +4,3 @@ export * from './job-model';
 export * from './locale-alternate-model';
 export * from './mail-model';
 export * from './private-config-model';
-export * from './site-map-url-model';
