--- functions/src/models/site-map-url-model.ts
@@ -1,14 +1,0 @@
-/**
- * Site Map Url Model
- * https://www.npmjs.com/package/sitemap
- */
-export class SiteMapUrlModel {
-    /** url of page */
-    url: string;
-    /** changefreq of page */
-    changefreq?: string;
-    /** priority of page */
-    priority?: number;
-    /** image url of page */
-    img?: string;
-}
