+++ package.json
@@ -34,7 +34,8 @@
     "test": "ng test",
     "test:cover": "del-cli \"./coverage\" && ng test --watch=false --code-coverage",
     "test:cover:open": "npm run test:cover && open-cli \"./coverage/index.html\"",
-    "test:cover:functions": "npm --prefix \"./functions\" run test:cover",
+    "/* Run twice to fix '%0 covered' issue */": "",
+    "test:cover:functions": "npm --prefix \"./functions\" run test:cover && npm --prefix \"./functions\" run test:cover",
     "test:cover:all": "npm run test:cover && npm run build:i18n && npm run server:i18n && npm run test:cover:functions && npm run test:combine:all",
     "test:combine:all": "del-cli \"./public/all-coverage\" && istanbul-combine -d \"./public/all-coverage\" -p summary -b \"./\" -r lcov -r html \"./coverage/coverage-final.json\" \"./functions/coverage/coverage-final.json\"",
     "lint:all": "npm run lint && npm run lint:functions",
@@ -78,59 +79,59 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "8.0.0",
-    "@angular/common": "8.0.0",
-    "@angular/compiler": "8.0.0",
-    "@angular/core": "8.0.0",
+    "@angular/animations": "8.1.2",
+    "@angular/common": "8.1.2",
+    "@angular/compiler": "8.1.2",
+    "@angular/core": "8.1.2",
     "@angular/fire": "5.2.1",
-    "@angular/forms": "8.0.0",
-    "@angular/platform-browser": "8.0.0",
-    "@angular/platform-browser-dynamic": "8.0.0",
-    "@angular/platform-server": "8.0.0",
-    "@angular/pwa": "0.800.1",
-    "@angular/router": "8.0.0",
-    "@angular/service-worker": "8.0.0",
-    "@fortawesome/fontawesome-free": "5.8.2",
-    "@ng-bootstrap/ng-bootstrap": "4.2.0",
+    "@angular/forms": "8.1.2",
+    "@angular/platform-browser": "8.1.2",
+    "@angular/platform-browser-dynamic": "8.1.2",
+    "@angular/platform-server": "8.1.2",
+    "@angular/pwa": "0.801.2",
+    "@angular/router": "8.1.2",
+    "@angular/service-worker": "8.1.2",
+    "@fortawesome/fontawesome-free": "5.9.0",
+    "@ng-bootstrap/ng-bootstrap": "5.1.0",
     "@ngx-loading-bar/core": "4.2.0",
     "@ngx-loading-bar/router": "4.2.0",
     "angular2-cookie-law": "^7.0.1",
     "angulartics2": "7.5.2",
     "bootstrap": "4.3.1",
     "classlist.js": "1.1.20150312",
-    "core-js": "3.1.3",
-    "firebase": "6.1.0",
-    "grpc": "1.21.1",
+    "core-js": "3.1.4",
+    "firebase": "6.3.1",
+    "grpc": "1.22.2",
     "protractor": "5.4.2",
     "rxjs": "6.5.2",
-    "tslib": "1.9.3",
+    "tslib": "1.10.0",
     "zone.js": "0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "0.800.1",
-    "@angular/cli": "8.0.1",
-    "@angular/compiler-cli": "8.0.0",
-    "@angular/language-service": "8.0.0",
-    "@commitlint/cli": "~8.0.0",
-    "@commitlint/config-conventional": "~8.0.0",
-    "@types/jasmine": "~3.3.13",
+    "@angular-devkit/build-angular": "0.801.2",
+    "@angular/cli": "8.1.2",
+    "@angular/compiler-cli": "8.1.2",
+    "@angular/language-service": "8.1.2",
+    "@commitlint/cli": "~8.1.0",
+    "@commitlint/config-conventional": "~8.1.0",
+    "@types/jasmine": "~3.3.14",
     "@types/jasminewd2": "~2.0.6",
-    "@types/node": "~12.0.4",
-    "acorn": "^6.1.1",
-    "angular-tslint-rules": "~1.14.0",
-    "browserslist": "~4.6.1",
+    "@types/node": "~12.6.8",
+    "acorn": "^6.2.0",
+    "angular-tslint-rules": "~1.18.0",
+    "browserslist": "~4.6.6",
     "browserstack-local": "~1.4.0",
     "codacy-coverage": "~3.4.0",
-    "codelyzer": "~5.0.1",
-    "coveralls": "~3.0.3",
-    "husky": "~2.3.0",
+    "codelyzer": "~5.1.0",
+    "coveralls": "~3.0.5",
+    "husky": "~3.0.1",
     "htmlhint": "~0.11.0",
     "istanbul-combine": "^0.3.0",
     "jasmine-core": "~3.4.0",
     "jasmine-spec-reporter": "~4.2.1",
     "jquery": "~3.4.1",
-    "karma": "~4.1.0",
-    "karma-chrome-launcher": "~2.2.0",
+    "karma": "~4.2.0",
+    "karma-chrome-launcher": "~3.0.0",
     "karma-cli": "~2.0.0",
     "karma-coverage-istanbul-reporter": "~2.0.5",
     "karma-jasmine": "~2.0.1",
@@ -141,19 +142,18 @@
     "node-sass": "~4.12.0",
     "popper.js": "~1.15.0",
     "remark-cli": "~6.0.1",
-    "remark-lint": "~6.0.4",
-    "remark-preset-lint-markdown-style-guide": "~2.1.2",
-    "remark-preset-lint-recommended": "^3.0.2",
-    "stylelint": "~10.0.1",
+    "remark-lint": "~6.0.5",
+    "remark-preset-lint-markdown-style-guide": "~2.1.3",
+    "remark-preset-lint-recommended": "^3.0.3",
+    "stylelint": "~10.1.0",
     "stylelint-config-recommended-scss": "~3.3.0",
     "stylelint-config-standard": "~18.3.0",
-    "stylelint-scss": "~3.8.0",
-    "tar": "^4.4.9",
-    "terser": "4.0.0",
-    "ts-node": "~8.2.0",
-    "tslint": "5.17.0",
+    "stylelint-scss": "~3.9.2",
+    "terser": "4.1.2",
+    "ts-node": "~8.3.0",
+    "tslint": "5.18.0",
     "typescript": "3.4.5",
-    "webpack": "~4.32.2",
+    "webpack": "~4.36.1",
     "webpack-bundle-analyzer": "~3.3.2"
   },
   "husky": {
