+++ src/app/app.module.ts
@@ -40,8 +40,10 @@ import { QuoteDetailComponent } from './pages/quote-detail/quote-detail.componen
 import { QuoteListComponent } from './pages/quote-list/quote-list.component';
 import { SearchComponent } from './pages/search/search.component';
 import { TaxonomyComponent } from './pages/taxonomy/taxonomy.component';
-import { AlertService, AuthService, CarouselService, ConfigService,
-    PagerService, PageService, PaginationService, SeoService } from './services';
+import {
+    AlertService, AuthService, CarouselService, ConfigService, JsonLDService,
+    PagerService, PageService, PaginationService, SeoService
+} from './services';
 import { ActiveTagsComponent } from './widgets/active-tags/active-tags.component';
 import { CustomHtmlComponent } from './widgets/custom-html/custom-html.component';
 import { LastJokesComponent } from './widgets/last-jokes/last-jokes.component';
@@ -108,6 +110,7 @@ import { SearchBarComponent } from './widgets/search-bar/search-bar.component';
     providers: [
         AlertService,
         SeoService,
+        JsonLDService,
         AuthService,
         PaginationService,
         PagerService,
