+++ src/app/services/index.ts
@@ -2,6 +2,7 @@ export * from './alert.service';
 export * from './auth.service';
 export * from './carousel.service';
 export * from './config.service';
+export * from './jsonld.service';
 export * from './page.service';
 export * from './pager.service';
 export * from './pagination.service';
