+++ src/app/testing/test.helper.module.spec.ts
@@ -18,8 +18,10 @@ import { PagerComponent } from '../components/pager/pager.component';
 import { SideBarComponent } from '../components/side-bar/side-bar.component';
 import { ScrollableDirective } from '../directives';
 import { NotFoundComponent } from '../pages/not-found/not-found.component';
-import { AlertService, AuthService, CarouselService, ConfigService,
-    PagerService, PageService, PaginationService, SeoService } from '../services';
+import {
+    AlertService, AuthService, CarouselService, ConfigService, JsonLDService,
+    PagerService, PageService, PaginationService, SeoService
+} from '../services';
 import { ActiveTagsComponent } from '../widgets/active-tags/active-tags.component';
 import { CustomHtmlComponent } from '../widgets/custom-html/custom-html.component';
 import { LastJokesComponent } from '../widgets/last-jokes/last-jokes.component';
@@ -73,6 +75,7 @@ export const activatedRouteStub = new ActivatedRouteStub();
     providers: [
         AlertService,
         SeoService,
+        JsonLDService,
         AuthService,
         PaginationService,
         PagerService,
